+++ explorer-nextjs/src/components/muiLink/index.tsx
@@ -1,6 +1,6 @@
-// Imports
 import { Link as MuiLink, type LinkProps as MuiLinkProps } from "@mui/material";
 import NextLink, { type LinkProps as NextLinkProps } from "next/link";
+// Imports
 import type { ReactNode } from "react";
 
 interface CustomLinkProps extends NextLinkProps, Omit<MuiLinkProps, "href"> {
