+++ explorer-nextjs/src/providers/ThemeProvider.tsx
@@ -1,8 +1,8 @@
 "use client";
-import { lightTheme } from "@/theme/theme";
 import { CssBaseline } from "@mui/material";
 import { ThemeProvider as MUIThemeProvider } from "@mui/material";
 import { AppRouterCacheProvider } from "@mui/material-nextjs/v15-appRouter";
+import { lightTheme } from "../theme/theme";
 
 const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
   return (
